feat: Removed email, sms and opene messages
