+++ frontend/src/contexts/app.context.tsx
@@ -3,12 +3,12 @@ import { ErrandsData, IErrand } from '@interfaces/errand';
 import { UiPhase } from '@interfaces/errand-phase';
 import { Notification as CaseDataNotification } from '@interfaces/notification';
 import { User } from '@interfaces/user';
-import { MessageNode } from '@services/casedata-message-service';
 import { Admin, emptyUser } from '@services/user-service';
 import { ReactNode, useState } from 'react';
 import { AppContext } from './app-context-interface';
 import { emptyErrandList } from '@services/casedata-errand-service';
 import { ErrandStatus } from '@interfaces/errand-status';
+import { MessageNode } from '@interfaces/message';
 
 export function AppWrapper({ children }: { children: ReactNode }) {
   const [isLoading, setIsLoading] = useState(false);
@@ -24,15 +24,12 @@ export function AppWrapper({ children }: { children: ReactNode }) {
   const [draftErrands, setDraftErrands] = useState<ErrandsData>(emptyErrandList);
   const [closedErrands, setClosedErrands] = useState<ErrandsData>(emptyErrandList);
   const [errand, setErrand] = useState<IErrand>({} as IErrand);
-  const [messages, setMessages] = useState<MessageNode[]>([]);
   const [conversation, setConversation] = useState<MessageNode[]>([]);
-  const [messageTree, setMessageTree] = useState<MessageNode[]>([]);
   const [assets, setAssets] = useState<Asset[]>([]);
   const [selectedErrandStatuses, setSelectedErrandStatuses] = useState<ErrandStatus[]>([ErrandStatus.ArendeInkommit]);
   const [municipalityId, setMunicipalityId] = useState<string>('');
   const [sidebarLabel, setSidebarLabel] = useState<string>('');
   const [administrators, setAdministrators] = useState<Admin[]>([]);
-  const [isCookieConsentOpen, setIsCookieConsentOpen] = useState(true);
   const [notifications, setNotifications] = useState<CaseDataNotification[]>([]);
   const [uiPhase, setUiPhase] = useState<UiPhase>(UiPhase.inkommet);
 
@@ -57,14 +54,8 @@ export function AppWrapper({ children }: { children: ReactNode }) {
         errand,
         setErrand: (errand: IErrand) => setErrand(errand),
 
-        messages,
-        setMessages: (messages: MessageNode[]) => setMessages(messages),
-
         conversation,
         setConversation: (conversation: MessageNode[]) => setConversation(conversation),
-
-        messageTree,
-        setMessageTree: (messages: MessageNode[]) => setMessageTree(messages),
 
         assets,
         setAssets: (assets: Asset[]) => setAssets(assets),
@@ -107,9 +98,6 @@ export function AppWrapper({ children }: { children: ReactNode }) {
         setAdministrators: (admins: Admin[]) => {
           setAdministrators(admins);
         },
-
-        isCookieConsentOpen,
-        setIsCookieConsentOpen: (isOpen: boolean) => setIsCookieConsentOpen(isOpen),
 
         uiPhase,
         setUiPhase: (phase: UiPhase) => setUiPhase(phase),
